WB4: collapse Effect workflow Services/Live ceremony
