--- src/ui/core/metadataSaveWorkflowLive.ts
@@ -1,43 +1,0 @@
-import { get } from 'svelte/store';
-
-import { tauriClient } from '../../lib/tauri/client';
-import { getCurrentFileList, persistPendingMetadataDraftsForCurrentSelection } from '../fileList';
-import { resetDirtyState } from '../metadataForm';
-import { clearPendingMetadataForFile, getPendingMetadataIntentEntries } from '../metadataState';
-import { metadataSaveInProgressStore } from '../metadataSaveState';
-import {
-	beginMetadataSaveInStatusPanel,
-	completeMetadataSaveInStatusPanel,
-	failMetadataSaveInStatusPanel,
-	initStatusPanel,
-	isStatusPanelProcessing,
-	pushStatusPanelTransientStatus,
-} from '../statusPanel';
-import {
-	makeMetadataSaveWorkflowServicesLayer,
-	type MetadataSaveWorkflowServices,
-} from './metadataSaveWorkflowServices';
-
-const liveMetadataSaveWorkflowServices: MetadataSaveWorkflowServices = {
-	getCurrentFileList,
-	initStatusPanel,
-	isStatusPanelProcessing,
-	pushStatusPanelTransientStatus,
-	isMetadataSaveInProgress: () => get(metadataSaveInProgressStore),
-	setMetadataSaveInProgress: (isInProgress) => {
-		metadataSaveInProgressStore.set(isInProgress);
-	},
-	persistPendingMetadataDraftsForCurrentSelection,
-	getPendingMetadataIntentEntries,
-	saveMetadataBatch: tauriClient.saveMetadataBatch,
-	clearPendingMetadataForFile,
-	resetDirtyState,
-	beginMetadataSaveInStatusPanel,
-	completeMetadataSaveInStatusPanel,
-	failMetadataSaveInStatusPanel,
-	console,
-};
-
-export const MetadataSaveWorkflowLive = makeMetadataSaveWorkflowServicesLayer(
-	liveMetadataSaveWorkflowServices,
-);
