+++ src/ui/outputPanel/AGENTS.md
@@ -11,7 +11,7 @@
   inside reactive markup (`$derived(...)`).
 
 ## Private Cluster
-- Files: `OutputPanelIsland.svelte`, `actions.ts`, `preview.ts`, `state.svelte.ts`, `outputPlanWorkflow.ts`, `outputPlanWorkflowLive.ts`, `outputPlanWorkflowServices.ts`, `__tests__/`.
+- Files: `OutputPanelIsland.svelte`, `actions.ts`, `preview.ts`, `state.svelte.ts`, `outputPlanWorkflow.ts`, `__tests__/`.
 - The cluster owns output directory, output naming, output-path preview, estimated-size display, tag preview display, collision review, and output-plan preflight.
 
 ## Allowed Agent Edits Without Escalation
